+++ src/cldflex/helpers.py
@@ -1,6 +1,6 @@
 import logging
 from slugify import slugify
-
+import pandas as pd
 
 log = logging.getLogger(__name__)
 
@@ -32,10 +32,16 @@ def deduplicate(unreliable_list):
 
 
 def delistify(df, sep):
-    for col in df.columns:
-        test_df = df[df[col].apply(lambda x: isinstance(x, list))]
-        if len(test_df) == 0:
-            continue
-        df[col] = df[col].fillna("").apply(list)
-        df[col] = df[col].apply(sep.join)
+    with pd.option_context("mode.chained_assignment", None):
+        for col in df.columns:
+            test_df = df[df[col].apply(lambda x: isinstance(x, list))]
+            if len(test_df) == 0:
+                continue
+            df[col] = df[col].fillna("").apply(list)
+            df[col] = df[col].apply(sep.join)
+    return df
+
+
+def listify(df, column, sep):
+    df[column] = df[column].apply(lambda x: x.split(sep))
     return df