+++ src/cldflex/lift2csv.py
@@ -7,16 +7,19 @@ import pandas as pd
 import yaml
 from bs4 import BeautifulSoup
 from slugify import slugify
-from cldflex.cldf import create_cldf
 from cldflex.cldf import create_dictionary_dataset
+from cldflex.cldf import create_rich_dataset
+from cldflex.cldf import create_wordlist_dataset
 from cldflex.helpers import add_to_list_in_dict
 from cldflex.helpers import deduplicate
 from cldflex.helpers import delistify
+from cldflex.helpers import listify
 
 
 log = logging.getLogger(__name__)
 
 
+# method for getting dictionary examples from entries
 def extract_examples(sense, dictionary_examples, sense_id):
     for ex_count, example in enumerate(sense.find_all("example")):
         example_dict = {"ID": f"{sense_id}-{ex_count}", "Sense_ID": sense_id}
@@ -52,82 +55,70 @@ def figure_out_gloss_language(entry):
 
 
 def parse_entries(entries):
-    parsed = []
-    morph_list = []
-    senses = []
-    dictionary_examples = []
+    parsed = []  # parsed entries
+    senses = []  # gathered senses
+    dictionary_examples = []  # gathered examples
     for entry in entries:
-        fields = {"Gramm": [], "Parameter_ID": []}
-        entry_id = entry["guid"]
-        fields["ID"] = entry_id
-        main_morph = {"ID": entry["guid"] + "-0", "Morpheme_ID": entry["guid"]}
-        for trait in entry.find_all("trait", recursive=False):
-            fields[trait["name"]] = trait["value"]
+        rec = {
+            "Gramm": [],
+            "Senses": [],
+        }  # the parsed entry, with different fields (i.e., CSV columns)
+        rec["ID"] = entry["guid"]
+        for trait in entry.find_all(
+            "trait", recursive=False
+        ):  # various traits, like morph-type
+            rec[trait["name"]] = trait["value"]
         for lexical_unit in entry.find_all("lexical-unit", recursive=False):
             for form in lexical_unit.find_all("form"):
-                fields["form_" + form["lang"]] = form.text
-                main_morph["form_" + form["lang"]] = form.text
+                rec["form_" + form["lang"]] = form.text
         for field in entry.find_all("field", recursive=False):
             for pseudoform in field.find_all("form"):
-                fields[
-                    slugify(field["type"]) + "_" + pseudoform["lang"]
-                ] = pseudoform.text
+                rec[slugify(field["type"]) + "_" + pseudoform["lang"]] = pseudoform.text
         for relation in entry.find_all("relation", recursive=False):
+            if "_" not in relation["ref"]:  # ignore relations without a reference
+                continue
             for trait in relation.find_all("trait"):
-                if "_" not in relation["ref"]:
-                    continue
                 add_to_list_in_dict(
-                    fields,
-                    f"""relation{relation["type"]}_{trait["name"]}_{slugify(trait["value"])}""",
+                    rec,
+                    f"relation{relation['type']}_{trait['name']}_{slugify(trait['value'])}",
                     relation["ref"].split("_")[1],
                 )
-
         for sense in entry.find_all("sense", recursive=False):
-            sense_id = sense["id"]  # todo human-readable option
-            sense_dict = {"ID": sense_id, "Entry_ID": entry_id}
-            fields["Parameter_ID"].append(sense["id"])
+            sense_id = sense["id"]  # todo: human-readable option
+            sense_dict = {"ID": sense_id, "Entry_ID": rec["ID"]}
+            rec["Senses"].append(sense["id"])
             for gramm in sense.find_all("grammatical-info"):
-                fields["Gramm"].append(gramm["value"])
+                rec["Gramm"].append(gramm["value"])
             for definition in sense.find_all("definition"):
                 for pseudoform in definition.find_all("form"):
-                    for x in [fields, sense_dict]:
+                    for x in [rec, sense_dict]:
                         add_to_list_in_dict(
                             x, "definition_" + pseudoform["lang"], pseudoform.text
                         )
             for gloss in sense.find_all("gloss"):
                 key = "gloss_" + gloss["lang"]
-                for x in [fields, main_morph, sense_dict]:
+                for x in [rec, sense_dict]:
                     add_to_list_in_dict(x, key, gloss.text)
             for note in sense.find_all("note"):
                 note_type = ("note_" + note.get("type", "")).strip("_")
                 for pseudoform in note.find_all("form"):
                     add_to_list_in_dict(
-                        fields, note_type + "_" + pseudoform["lang"], pseudoform.text
+                        rec, note_type + "_" + pseudoform["lang"], pseudoform.text
                     )
             for reversal in sense.find_all("reversal"):
                 for form in reversal.find_all("form"):
-                    add_to_list_in_dict(fields, "reversal_" + form["lang"], form.text)
+                    add_to_list_in_dict(rec, "reversal_" + form["lang"], form.text)
             senses.append(sense_dict)
             extract_examples(sense, dictionary_examples, sense_id)
 
-        main_morph["morph-type"] = fields.get("morph-type", "?")
-        main_morph["Parameter_ID"] = fields.get("Parameter_ID", "na")
-        morph_list.append(main_morph)
         for i, allomorph in enumerate(entry.find_all("variant", recursive=False)):
             form = allomorph.find("form")
-            add_to_list_in_dict(fields, "variant_" + form["lang"], form.text)
-            new_morph = main_morph.copy()
-            new_morph["ID"] = f"""{entry["guid"]}-{i+1}"""
-            new_morph["Parameter_ID"] = fields.get("Parameter_ID", "na")
-            new_morph["morph-type"] = allomorph.select("trait[name='morph-type']")[0][
-                "value"
-            ]
-            for form in allomorph.find_all("form"):
-                new_morph["form_" + form["lang"]] = form.text
-            morph_list.append(new_morph)
-        fields["Gramm"] = deduplicate(fields["Gramm"])
-        parsed.append(fields)
-    return parsed, morph_list, senses, dictionary_examples
+            for trait in allomorph.find_all("trait", recursive=False):
+                add_to_list_in_dict(rec, "variant_" + trait["name"], trait["value"])
+            add_to_list_in_dict(rec, "variant_" + form["lang"], form.text)
+        rec["Gramm"] = deduplicate(rec["Gramm"])
+        parsed.append(rec)
+    return parsed, senses, dictionary_examples
 
 
 def convert(
@@ -142,47 +133,57 @@ def convert(
     elif not conf:
         with open(config_file, encoding="utf-8") as f:
             conf = yaml.safe_load(f)
-    sep = conf.get("csv_cell_separator", "; ")
-    obj_lg = conf.get("obj_lg", None)
-    gloss_lg = conf.get("gloss_lg", None)
+    sep = conf.get(
+        "csv_cell_separator", "; "
+    )  # separator used in cells with multiple values
+
     log.info(f"Parsing {lift_file.resolve()}")
     with open(lift_file, "r", encoding="utf-8") as f:
         lexicon = BeautifulSoup(f.read(), features="xml")
 
-    for entry in lexicon.find_all("entry"):
+    obj_lg = conf.get("obj_lg", None)  # main object language
+    gloss_lg = conf.get("gloss_lg", None)  # main gloss language
+
+    for entry in lexicon.find_all(
+        "entry"
+    ):  # if not defined, they are deducted from the data
         if not gloss_lg:
             gloss_lg = figure_out_gloss_language(entry)
         if not obj_lg:
             obj_lg = entry.find("form")["lang"]
 
-    obj_key = f"form_{obj_lg}"
-    definition_key = f"definition_{gloss_lg}"
-    gloss_key = f"gloss_{gloss_lg}"
+    obj_key = f"form_{obj_lg}"  # <form lang="X"><text>Y</text></form> becomes form_X: Y
+    definition_key = f"definition_{gloss_lg}"  # <definition><form lang="X"><text>Y</text></form></definition> becomes defition_X: Y
+    gloss_key = (
+        f"gloss_{gloss_lg}"  # <gloss lang="X"><text>Y</text></gloss> becomes X_gloss: Y
+    )
     var_key = "variant_" + obj_lg
 
     var_dict = {}
-    entries, morph_list, senses, dictionary_examples = parse_entries(
-        lexicon.find_all("entry")
-    )
+    entries, senses, dictionary_examples = parse_entries(lexicon.find_all("entry"))
     entries = pd.DataFrame.from_dict(entries)
     senses = pd.DataFrame.from_dict(senses)
+
     for key in [definition_key, gloss_key]:
         if key not in senses.columns:
             senses[key] = np.nan
+    # fill sense descriptions with glosses
     senses["Description"] = senses.apply(
         lambda x: x[definition_key]
         if not pd.isnull(x[definition_key])
         else x[gloss_key],
         axis=1,
     )
+    # fill sense names with glosses and definitions
+
     senses["Name"] = senses.apply(
         lambda x: " / ".join(x[gloss_key])
-        if not pd.isnull(x[gloss_key])
+        if not pd.isnull(x[gloss_key]).all()
         else x[definition_key],
         axis=1,
     )
-    unmodified_entries = entries.copy()
 
+    # method for printing entries in log
     def entry_repr(entry_id):
         entry = entries[entries["ID"] == entry_id].iloc[0]
         meanings = entry.get(gloss_key, entry.get(definition_key, ""))
@@ -192,16 +193,35 @@ def convert(
             ggg = "unknown meaning"
         else:
             ggg = " / ".join(meanings)
-        if isinstance(entry.get(obj_key, None), list):
-            form_str = " / ".join(entry[obj_key])
+        if isinstance(entry.get("Form", None), list):
+            form_str = " / ".join(entry["Form"])
         else:
-            form_str = entry.get(obj_key, "MISSING FORM")
-        return f"""{form_str} '{ggg}' ({','.join(entry["Gramm"])}, {entry["morph-type"]}) [{entry["ID"]}]"""
+            form_str = entry.get("Form", "MISSING FORM")
+        return f"""{form_str} '{ggg}' ({','.join(entry["Gramm"])}, {entry["Type"]})"""
 
+    entries.rename(
+        columns={
+            obj_key: "Form",
+            var_key: "Variants",
+            "Senses": "Parameter_ID",
+            gloss_key: "Gloss",
+            "morph-type": "Type",
+        },
+        inplace=True,
+    )
+    gloss_key = "Gloss"
+    entries["Variants"] = entries["Variants"].apply(
+        lambda d: d if isinstance(d, list) else []
+    )
+    entries["Language_ID"] = obj_lg
+
+    # 1. find all variants
+    # 2. compile dictionary mapping entry IDs to variants
     for col in entries.columns:
         if "variant-type" not in col:
             continue
         variants = entries[~(pd.isnull(entries[col]))]
+        log.info(f"Parsing variants of type '{col} ({len(variants)} found)")
         for entry in variants.to_dict("records"):
             if len(entry[col]) > 1:
                 msg = f"""The entry {entry_repr(entry["ID"])} is stored as a variant ({col}) of multiple main entries:"""
@@ -211,33 +231,24 @@ def convert(
             for entry_id in entry[col]:
                 add_to_list_in_dict(var_dict, entry_id, entry)
 
-    check_variants = var_key in entries.columns
-
-    entries["Name"] = entries[obj_key]
-    entries[obj_key] = entries[obj_key].apply(
-        lambda x: [x] if isinstance(x, str) else []
-    )
-    for key in [gloss_key, definition_key]:
+    # listify gloss and variant columns
+    for key in [gloss_key, var_key, "variant_morph-type"]:
         if key in entries:
             entries[key] = entries[key].apply(
                 lambda x: [] if not isinstance(x, list) else x
             )
 
-    def process_variant(entry, variant, new_variant_morphs, var_dict, var_count, i):
-        if variant["ID"] in var_dict:
+    entry_variants = {}
+
+    def process_variant(entry, variant, var_count, idx):
+        if variant["ID"] in var_dict:  # in theory, variants can have other variants
             for varvariant in var_dict[variant["ID"]]:
                 log.warning(
-                    f"""The variant {entry_repr(variant["ID"])} of the entry {entry_repr(entry["ID"])} has the subvariant {entry_repr(varvariant["ID"])}. Is this accurate?"""
+                    f"The variant {entry_repr(variant['ID'])} of the entry {entry_repr(entry['ID'])} has the subvariant {entry_repr(varvariant['ID'])}. Is this accurate?"
                 )
-                varvariant = varvariant.copy()
                 var_count += 1
-                if var_count < 10:
-                    process_variant(
-                        entry, varvariant, new_variant_morphs, var_dict, var_count, i
-                    )
-        log.debug(
-            f"""Adding variant {variant["ID"]} {variant[obj_key]} to entry {entry_repr(entry["ID"])}"""
-        )
+                if var_count < 10:  # stop recursion at some point
+                    process_variant(entry, varvariant.copy(), var_count, idx)
         if variant["Gramm"] and variant["Gramm"] != entry["Gramm"]:
             log.warning(
                 f"""The entry {entry_repr(variant["ID"])} is stored as a variant of {entry_repr(entry["ID"])}. It will not retain its part of speech."""
@@ -249,68 +260,68 @@ def convert(
                     f"""The entry {entry_repr(entry["ID"])} is stored as having a different meaning than its variant {entry_repr(variant["ID"])}"""
                 )
         else:
-            log.debug(
-                f"""inheriting gloss from {entry_repr(entry["ID"])} for {entry_repr(variant["ID"])}"""
-            )
             variant[gloss_key] = entry.get(gloss_key, entry.get(definition_key, ""))
         variant["Parameter_ID"] = entry["Parameter_ID"]
-        entry[obj_key].append(variant[obj_key])
-        variant["Variant_ID"] = variant["ID"]
-        variant["Morpheme_ID"] = entry["ID"]
-        variant["ID"] = f"""{entry["ID"]}-{var_count+i}"""
-        new_variant_morphs.append(variant)
+        add_to_list_in_dict(entry_variants, entry["ID"], variant)
 
-    morphs = pd.DataFrame.from_dict(morph_list)
+    def resolve_variants(entry):
+        for idx, variant in enumerate(  # iterate gathered variants for this entry
+            var_dict.get(entry["ID"], [])
+        ):
+            process_variant(
+                entry=entry, variant=variant, var_count=len(entry["Variants"]), idx=idx
+            )
 
-    new_variant_morphs = []
-    for entry in entries.to_dict("records"):
-        var_count = 1
-        if check_variants and isinstance(entry[var_key], list):
-            for variant in entry[var_key]:
-                var_count += 1
-                entry[obj_key].append(variant)
-        if entry["ID"] in var_dict:
-            for i, variant in enumerate(var_dict[entry["ID"]]):
-                variant = variant.copy()
-                process_variant(
-                    entry, variant, new_variant_morphs, var_dict, var_count, i
-                )
+    entries.apply(lambda x: resolve_variants(x), axis=1)
 
-    new_variant_morphs = pd.DataFrame.from_dict(new_variant_morphs)
-
-    if len(new_variant_morphs) > 0:
-        morphs = morphs[~(morphs["Morpheme_ID"].isin(new_variant_morphs["Variant_ID"]))]
-        morphs = pd.concat([morphs, new_variant_morphs])
-    morphs["Language_ID"] = obj_lg
-    morphs.rename(
-        columns={obj_key: "Form", gloss_key: "Gloss", "morph-type": "Type"},
-        inplace=True,
-    )
-
+    # delete variants
     for col in entries.columns:
         if "variant-type" not in col:
             continue
         variants = entries[~(pd.isnull(entries[col]))]
         entries = entries.loc[~(entries.index.isin(variants.index))]
 
+    # split up entries into lexemes, stems, morphemes, and morphs
+    morphemes = entries[~(entries["Type"].isin(["phrase", "stem"]))]
+    lexemes = entries[(entries["Type"].isin(["root", "stem"]))]
 
-    entries[obj_key] = entries[obj_key].apply(sorted)
-    entries[obj_key] = entries[obj_key].apply(deduplicate)
-    entries = delistify(entries, sep)
-    entries["Language_ID"] = obj_lg
+    def split_into_variants(rec, abstract_key, main_variant=None):
+        for idx, (form, morph_type) in enumerate(
+            zip(
+                [rec["Form"]] + rec["Variants"],
+                [rec["Type"]] + rec["variant_morph-type"],
+            )
+        ):
+            new_rec = rec.copy()
+            new_rec[abstract_key] = rec["ID"]
+            new_rec["ID"] = rec["ID"] + f"-{idx}"
+            new_rec["Form"] = form
+            new_rec["Type"] = morph_type
+            if main_variant:
+                new_rec[main_variant] = rec["ID"] + "-0"
+            yield dict(new_rec)
+        for external_variant in entry_variants.get(rec["ID"], []):
+            external_variant[abstract_key] = rec["ID"]
+            if main_variant:
+                external_variant[main_variant] = rec["ID"] + "-0"
+            yield external_variant
 
-    morphemes = entries.copy()
-    morphemes = morphemes[(~(morphemes["morph-type"].isin(["phrase"])))]
-    morphemes.rename(
-        columns={
-            gloss_key: "Gloss",
-            definition_key: "Meaning",
-            obj_key: "Form",
-            "morph-type": "Type",
-        },
-        inplace=True,
+    morphs = []
+    stems = []
+    morphemes.apply(
+        lambda x: morphs.extend(split_into_variants(x, "Morpheme_ID")), axis=1
     )
-    morphs = morphs[(morphs["Morpheme_ID"].isin(list(morphemes["ID"])))]
+
+    lexemes.apply(
+        lambda x: stems.extend(
+            split_into_variants(x, "Lexeme_ID", main_variant="Main_Stem")
+        ),
+        axis=1,
+    )
+
+    morphs = pd.DataFrame.from_dict(morphs)
+    stems = pd.DataFrame.from_dict(stems)
+    stems = stems[(stems["Type"].isin(["root", "stem"]))]
 
     sentence_path = Path(output_dir / "sentences.csv")
     ref_pattern = re.compile(r"^(\d+.\d)+$")
@@ -325,9 +336,7 @@ def convert(
             juicy_columns = ["Analyzed_Word", "Gloss"]
             glossed_examples.dropna(subset=juicy_columns, inplace=True)
             for col in juicy_columns:
-                glossed_examples[col] = glossed_examples[col].apply(
-                    lambda x: x.split("\t")
-                )
+                glossed_examples = listify(glossed_examples, col, "\t")
             enriched_examples = []
             for ex in dictionary_examples:
                 successful = False
@@ -370,54 +379,44 @@ def convert(
     glottocode = conf.get("Glottocode", conf.get("Language_ID", None))
     with pd.option_context("mode.chained_assignment", None):
         if glottocode:
-            for df in [unmodified_entries, morphemes, morphs, dictionary_examples]:
+            for df in [entries, morphemes, morphs, dictionary_examples]:
                 df["Language_ID"] = glottocode
         else:
-            for df in [unmodified_entries, morphemes, morphs, dictionary_examples]:
+            for df in [entries, morphemes, morphs, dictionary_examples]:
                 df["Language_ID"] = obj_lg
 
-    morphemes.fillna("", inplace=True)
-    morphemes.to_csv(output_dir / "morphemes.csv", index=False)
-    delistify(senses, sep)
-    senses.to_csv(output_dir / "senses.csv", index=False)
-    delistify(morphs, sep)
-    morphs.drop_duplicates(
-        subset=[x for x in morphs.columns if x != "ID"], inplace=True
-    )  # entries may be "variants" of other entries in multiple ways
+    for df, name in [
+        (entries, "entries"),
+        (stems, "stems"),
+        (lexemes, "lexemes"),
+        (morphs, "morphs"),
+        (morphemes, "morphemes"),
+        (senses, "senses"),
+    ]:
+        df = delistify(df, sep)
+        df.to_csv(output_dir / f"{name}.csv", index=False)
 
-    morphs.to_csv(output_dir / "morphs.csv", index=False)
-
-    unmodified_entries["Name"] = unmodified_entries[obj_key]
-    delistify(unmodified_entries, sep)
-    unmodified_entries.to_csv(output_dir / "entries.csv", index=False)
     if cldf:
         cldf_settings = conf.get("cldf", {})
         metadata = cldf_settings.get("metadata", {})
-        if cldf_settings.get("lexicon", None) == "wordlist":
-            unmodified_entries.rename(
-                columns={gloss_key: "Meaning", obj_key: "Form"}, inplace=True
-            )
-            morphemes["Form"] = morphemes["Form"].apply(lambda x: x.split(sep))
-            morphemes["Parameter_ID"] = morphemes["Parameter_ID"].apply(
-                lambda x: x.split(sep)
-            )
-            tables = {
-                "FormTable": morphemes,
-                "ParameterTable": senses,
-            }
-            create_cldf(
-                tables=tables,
+        mode = cldf_settings.get("lexicon", "dictionary")
+        if mode == "wordlist":
+            create_wordlist_dataset(
+                forms=entries,
+                senses=senses,
                 glottocode=glottocode,
                 metadata=metadata,
                 output_dir=output_dir,
                 cwd=lift_file.parents[0],
                 sep=sep,
+                parameters=cldf_settings.get("parameters", "multi"),
             )
-        else:
+        elif mode == "dictionary":
             if cldf_settings.get("drop_empty", False):
                 senses = senses[senses["Description"] != ""]
+            senses = senses[senses["Entry_ID"].isin(entries["ID"].values)]
             create_dictionary_dataset(
-                unmodified_entries,
+                entries,
                 senses,
                 metadata=metadata,
                 examples=dictionary_examples,
@@ -425,5 +424,31 @@ def convert(
                 output_dir=output_dir,
                 cwd=lift_file.parents[0],
             )
+        elif mode == "rich":
+            tables = {}
+            with pd.option_context("mode.chained_assignment", None):
+                for namedf in [morphs, lexemes, morphemes, stems]:
+                    namedf.rename(columns={"Form": "Name"}, inplace=True)
+
+            for name, df in [
+                ("morphemes", morphemes),
+                ("morphs", morphs),
+                ("lexemes", lexemes),
+                ("stems", stems),
+                ("senses", senses),
+            ]:
+                if len(df) > 0:
+                    tables[name] = df
+            create_rich_dataset(
+                tables=tables,
+                glottocode=glottocode,
+                metadata=metadata,
+                output_dir=output_dir,
+                cwd=lift_file.parents[0],
+                sep=sep,
+                parameters=cldf_settings.get("parameters", "multi"),
+            )
+        else:
+            raise ValueError(mode)
 
     return morphs